backend/api-gateway/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
requests
pyjwt
bcrypt
